refactor(exec): Extract more common types
