+++ lib/manager/poetry/artifacts.ts
@@ -1,5 +1,5 @@
 import is from '@sindresorhus/is';
-import { exec, ExecOptions } from '../../util/exec';
+import { exec } from '../../util/exec';
 import { logger } from '../../logger';
 import { UpdateArtifact, UpdateArtifactsResult } from '../common';
 import {
@@ -8,6 +8,7 @@ import {
   readLocalFile,
   writeLocalFile,
 } from '../../util/fs';
+import { ExecOptions } from '../../util/exec/common';
 
 export async function updateArtifacts({
   packageFileName,
