+++ lib/util/exec/common.ts
@@ -1,3 +1,9 @@
+import {
+  exec as cpExec,
+  ExecOptions as ChildProcessExecOptions,
+} from 'child_process';
+import { promisify } from 'util';
+
 export type Opt<T> = T | null | undefined;
 
 export enum BinarySource {
@@ -23,3 +29,25 @@ export interface DockerOptions {
   envVars?: Opt<Opt<string>[]>;
   cwd?: Opt<string>;
 }
+
+export interface RawExecOptions extends ChildProcessExecOptions {
+  encoding: string;
+}
+
+export type ExtraEnv<T = unknown> = Record<string, T>;
+
+export interface ExecOptions extends ChildProcessExecOptions {
+  subDirectory?: string;
+  extraEnv?: Opt<ExtraEnv>;
+  docker?: Opt<DockerOptions>;
+}
+
+export interface ExecResult {
+  stdout: string;
+  stderr: string;
+}
+
+export const rawExec: (
+  cmd: string,
+  opts: RawExecOptions
+) => Promise<ExecResult> = promisify(cpExec);
